+++ test-corpus/test_corpus.py
@@ -24,6 +24,7 @@ from schemathesis.core.errors import (
     format_exception,
 )
 from schemathesis.core.failures import Failure
+from schemathesis.core.jsonschema import BundleError
 from schemathesis.core.result import Ok
 from schemathesis.core.transport import Response
 from schemathesis.engine import Status, events, from_schema
@@ -378,7 +379,7 @@ def test_default(corpus, filename, tmp_path):
     schema = _load_schema(corpus, filename)
     try:
         schema.as_state_machine()()
-    except (RefResolutionError, IncorrectUsage, LoaderError, InvalidSchema, InvalidStateMachine):
+    except (RefResolutionError, IncorrectUsage, LoaderError, InvalidSchema, InvalidStateMachine, BundleError):
         pass
 
     schema.config.update(suppress_health_check=list(HealthCheck))
@@ -490,6 +491,8 @@ def should_ignore_error(schema_id: str, event: events.NonFatalError) -> bool:
         return True
     if "Malformed path template" in formatted:
         return True
+    if "Unknown type:" in formatted:
+        return True
     if "Unresolvable JSON pointer" in formatted:
         return True
     if "Ensure that the definition complies with the OpenAPI specification" in formatted:
@@ -499,6 +502,8 @@ def should_ignore_error(schema_id: str, event: events.NonFatalError) -> bool:
     if "is not defined in API operation" in formatted:
         return True
     if "contain invalid link definitions" in formatted:
+        return True
+    if "Cannot bundle" in formatted:
         return True
     if RECURSIVE_REFERENCE_ERROR_MESSAGE in formatted:
         return True