+++ test/test_dereferencing.py
@@ -534,7 +534,7 @@ def test_(request, case):
     assert case.path == "/users"
     assert case.method == "POST"
     if not hasattr(case.meta.phase.data, "description"):
-        assert case.body is None or case.body == 0
+        assert case.body in (None, 0, 1)
 """,
         paths={
             "/users": {
@@ -869,3 +869,50 @@ def test_missing_file_in_resolution(ctx, testdir, cli, snapshot_cli, openapi3_ba
     raw_schema_path.write_text(json.dumps(schema), "utf8")
 
     assert cli.run(str(raw_schema_path), f"--url={openapi3_base_url}") == snapshot_cli
+
+
+def test_unresolvable_operation(ctx, cli, snapshot_cli, openapi3_base_url):
+    schema_path = ctx.openapi.write_schema(
+        {
+            "/test": {
+                "post": {
+                    "$ref": "#/0",
+                    "responses": {
+                        "default": {
+                            "description": "Ok",
+                        }
+                    },
+                }
+            }
+        }
+    )
+    assert cli.run(str(schema_path), f"--url={openapi3_base_url}", "--phases=fuzzing") == snapshot_cli
+
+
+def test_iter_when_ref_resolves_to_none_in_body(ctx):
+    # Key0 -> Key1 -> Key2 -> Key3 (None)
+    schema = ctx.openapi.build_schema(
+        {
+            "/test": {
+                "get": {
+                    "requestBody": {
+                        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Key0"}}},
+                        "required": True,
+                    },
+                    "responses": {"200": {"description": "OK"}},
+                }
+            }
+        },
+        components={
+            "schemas": {
+                **{f"Key{idx}": {"$ref": f"#/components/schemas/Key{idx + 1}"} for idx in range(3)},
+                "Key3": None,
+            }
+        },
+    )
+
+    schema = schemathesis.openapi.from_dict(schema)
+
+    # Should not fail
+    for _ in schema.get_all_operations():
+        pass