+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 ### :bug: Fixed
 
 - Properly generate boundary values for negative `maximum` during the coverage phase.
+- **CLI**: Internal error if referenced API operation is unresolvable.
+
+### :wrench: Changed
+
+- Improve error reporting on incorrect `type` values.
 
 ### :racing_car: Performance